Renamespace compose-mode components
