+++ osu.Game.Tests/Visual/TestCaseHitObjectComposer.cs
@@ -17,7 +17,7 @@ using osu.Game.Rulesets.Osu.Edit.Masks.HitCircleMasks;
 using osu.Game.Rulesets.Osu.Edit.Masks.HitCircleMasks.Components;
 using osu.Game.Rulesets.Osu.Objects;
 using osu.Game.Screens.Edit.Screens.Compose;
-using osu.Game.Screens.Edit.Screens.Compose.Layers;
+using osu.Game.Screens.Edit.Screens.Compose.Components;
 using osu.Game.Tests.Beatmaps;
 
 namespace osu.Game.Tests.Visual
