+++ osu.Game.Tests/Visual/TestCaseZoomableScrollContainer.cs
@@ -10,7 +10,7 @@ using osu.Framework.Graphics.Shapes;
 using osu.Framework.MathUtils;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Cursor;
-using osu.Game.Screens.Edit.Screens.Compose.Timeline;
+using osu.Game.Screens.Edit.Screens.Compose.Components.Timeline;
 using OpenTK;
 using OpenTK.Graphics;
